requirements.txt
streamlit
torch
pandas
numpy
matplotlib
seaborn
ultralytics
opencv-python
pillow
requests
